fix(maskfile): use git status for changes check
`git diff` only checks for changes in files known to git which means if
using it to check if there are any changes in theme directory, it won't
detect anything the first time when theme hasn't been committed yet.

Use `git status` for the check as it outputs lines for untracked files
as well, which will allow us to detect this case as well.